rangeproof: remove rewinding ability to access the "extra" 64 bytes of sidechannel space
This space has never been usable, in the sense that rangeproof_sign would
refuse to use it, and it complicates the rewinding logic a fair bit to
retain the ability of the rewinder to access it.

This does result in a minor API change, which is that the returned `m_len`
variable, which indicates the total size of the rangeproof sidechannel,
is reduced by 64 bytes for any proof that covers a multiple-of-4 number
of bit. This change is reflected as a change in the unit tests.