Avoid using out-of-bounds field elements (in impossible cases)
secp256k1_fe_set_b32_limit says that when it returns 0, one is not allowed to use
the resulting output value.

This refactors the code so that the existing value of t is maintained in cases where
sha256 would output an out-of-bounds hash value.

Note: This situation is cryptographically impossible to occur.