shallue_van_de_woestijne support for t = 0
The current implementation returns an off-curve point for the input t=0.
This is due to performing a "0/0" computation.

Here we change the code to explicitly return the on-curve point (d, sqrt(8)),
which is the point that the paper Indifferentiable Hashing to Barreto–Naehrig
Curves suggests returning in this case.

Note: At the moment it is cryptographically impossible for the input t to be 0.