Avoid using out-of-bounds field elements (in impossible cases)
secp256k1_fe_set_b32_limit says that when it returns 0, one is not allowed to use
the resulting output value.

This refactors the code so that 0 is returned early (indicating failure) in cases
where sha256 would output an out-of-bounds hash value. This makes
secp256k1_generator_generate_internal variable-time in its "t" argument,
but this not a problem because this value is public in applications.

Note: This situation is cryptographically impossible to occur.

Co-authored-by: Russell O'Connor <roconnor@blockstream.io>